Move tests into multiple files
